[Feature] auto select file contents after opening it.
After opening the file, the first thing a user will do is select
all of it's content to copy it. Why not we do it for them!

We would not like to copy the content to clipboard automatically
as it would create unwanted confusion and frustration if someone
clicks on an image file in sidebar by mistake.